# Introduced a select cursor.