The goal of the PR is more to allow in-place select to unify code in the range query.
There seems to be a small perf improvment as a bonus.

Before
```
test column_index::optional_index::tests::bench::bench_translate_codec_to_orig_10percent_filled_0comma005percent_hit ... bench:       1,219 ns/iter (+/- 25)
test column_index::optional_index::tests::bench::bench_translate_codec_to_orig_1percent_filled_0comma005percent_hit  ... bench:          18 ns/iter (+/- 0)
test column_index::optional_index::tests::bench::bench_translate_codec_to_orig_1percent_filled_10percent_hit         ... bench:       4,004 ns/iter (+/- 81)
test column_index::optional_index::tests::bench::bench_translate_codec_to_orig_1percent_filled_full_scan             ... bench:      38,154 ns/iter (+/- 782)
test column_index::optional_index::tests::bench::bench_translate_codec_to_orig_90percent_filled_0comma005percent_hit ... bench:       7,296 ns/iter (+/- 134)
test column_index::optional_index::tests::bench::bench_translate_codec_to_orig_90percent_filled_full_scan            ... bench:  17,812,681 ns/iter (+/- 57,394)
```

After
```
test column_index::optional_index::tests::bench::bench_translate_codec_to_orig_10percent_filled_0comma005percent_hit ... bench:       1,140 ns/iter (+/- 19)
test column_index::optional_index::tests::bench::bench_translate_codec_to_orig_1percent_filled_0comma005percent_hit  ... bench:           2 ns/iter (+/- 0)
test column_index::optional_index::tests::bench::bench_translate_codec_to_orig_1percent_filled_10percent_hit         ... bench:       2,373 ns/iter (+/- 38)
test column_index::optional_index::tests::bench::bench_translate_codec_to_orig_1percent_filled_full_scan             ... bench:      22,787 ns/iter (+/- 225)
test column_index::optional_index::tests::bench::bench_translate_codec_to_orig_90percent_filled_0comma005percent_hit ... bench:       7,120 ns/iter (+/- 96)
test column_index::optional_index::tests::bench::bench_translate_codec_to_orig_90percent_filled_full_scan            ... bench:  17,150,903 ns/iter (+/- 36,651)

```